+++ packages/create-sitecore-jss/src/templates/nextjs-styleguide/src/lib/next-config/plugins/disconnected.js
@@ -13,8 +13,10 @@ const disconnectedPlugin = (nextConfig = {}) => {
 
   return Object.assign({}, nextConfig, {
     async rewrites() {
+      const existingRewrites = await (nextConfig.rewrites ? nextConfig.rewrites() : []);
+
       // When disconnected we proxy to the local faux layout service host, see scripts/disconnected-mode-proxy.ts
-      return [
+      const disconnectedModeRewrites = [
         // API endpoints
         {
           source: '/sitecore/api/:path*',
@@ -26,6 +28,16 @@ const disconnectedPlugin = (nextConfig = {}) => {
           destination: `${disconnectedServerUrl}/data/media/:path*`,
         },
       ];
+
+      // Remove conflicting rewrites provided in a base next.config
+      return existingRewrites
+        .filter(
+          (rewrite) =>
+            !disconnectedModeRewrites.find(
+              (disconnectedRewrite) => disconnectedRewrite.source === rewrite.source
+            )
+        )
+        .concat(disconnectedModeRewrites);
     },
     webpack: (config, options) => {
       // Prevent webpack-5 from throwing error for sitecore-import.json when app first starts